Improve NCE and Finish Up Refactor
This commit mainly finishes up refactor by fixing everything brought up in the CR + Improving NCE somewhat and actually killing the child processes properly now.

+++ app/src/main/cpp/skyline/nce/guest.h
@@ -42,6 +42,6 @@ namespace skyline {
          * @param pc The address of PC when the call was being done
          * @param svc The SVC ID of the SVC being called
          */
-        void SvcHandler(u64 pc, u32 svc);
+        void SvcHandler(u64 pc, u16 svc);
     }
 }
